refactor: params
