+++ index.js
@@ -1,7 +1,7 @@
 const chokidar = require('chokidar')
 const { Server } = require('ws')
 const StaticServer = require('./lib/server')
-const trigger = require('./lib/trigger')
+const defaultSubscribe = require('./lib/subscribe')
 const defaultParams = require('./lib/params')
 
 module.exports = class extends StaticServer {
@@ -13,24 +13,16 @@ module.exports = class extends StaticServer {
 
   initialize() {
     this.timer = null
-    this.clients = []
     this.server = null
     this.ws = null
     this.watcher = null
-    this.current = {}
-    this.trigger = trigger
+    this.clients = []
+    this.$subscribe = defaultSubscribe
   }
 
-  set listener(fn) {
-    if (typeof fn === 'function') {
-      this.trigger = fn
-    }
-  }
-
-  get status() {
-    return {
-      ...this.current,
-      running: !!this.server,
+  set subscribe($subscribe) {
+    if (typeof $subscribe === 'function') {
+      this.$subscribe = $subscribe
     }
   }
 
@@ -43,20 +35,12 @@ module.exports = class extends StaticServer {
 
   removeClient(client) {
     this.clients = this.clients.filter(c => c !== client)
-    this.trigger({
-      ...defaultParams,
-      event: 'info',
-      message: `clients: ${this.clients.length}`,
-    })
+    this.$subscribe({ ...defaultParams, clients: this.clients.length })
   }
 
   start(port = 2333) {
     if (this.server) {
-      this.trigger({
-        ...defaultParams,
-        event: 'info',
-        message: 'The server is running...',
-      })
+      this.$subscribe({ ...defaultParams, status: 'running' })
       return
     }
 
@@ -70,17 +54,11 @@ module.exports = class extends StaticServer {
     this.watcher.on('all', (event, filePath) => {
       clearTimeout(this.timer)
       this.timer = setTimeout(() => {
-        this.current = { event, path: filePath }
-        this.trigger({
+        this.$subscribe({
+          ...defaultParams,
           event,
           path: filePath,
-          message: '',
-          reloadCss: () => {
-            this.clients.forEach(client => client.send('css'))
-          },
-          reloadPage: () => {
-            this.clients.forEach(client => client.send('reload'))
-          },
+          trigger: type => this.clients.forEach(client => client.send(type)),
         })
       }, 100)
     })
@@ -90,17 +68,9 @@ module.exports = class extends StaticServer {
       client.on('close', () => this.removeClient(client))
       client.on('error', () => this.removeClient(client))
       this.clients.push(client)
-      this.trigger({
-        ...defaultParams,
-        event: 'info',
-        message: `clients: ${this.clients.length}`,
-      })
+      this.$subscribe({ ...defaultParams, clients: this.clients.length })
     })
 
-    this.trigger({
-      ...defaultParams,
-      event: 'info',
-      message: `Server running: http://127.0.0.1:${port}\n  CTRL + C to shutdown`,
-    })
+    this.trigger({ ...defaultParams, port, status: 'start' })
   }
 }
