+++ lib/inject.js
@@ -13,7 +13,7 @@
     ws.onmessage = function(e) {
       e = e.data
       console.log(e)
-      if (e === 'reload') {
+      if (e === 'page') {
         location.reload()
       }
       if (e === 'css') {
