+++ lib/params.js
@@ -1,7 +1,8 @@
 module.exports = {
-  event: '',
-  message: '',
-  path: '',
-  reloadCss: () => null,
-  reloadPage: () => null,
+  event: undefined,
+  path: undefined,
+  port: undefined,
+  clients: 0,
+  status: 'stop',
+  trigger: () => null,
 }
